Move TeensyRow logic to seperate component fixing multiple dialog trigger issue
